refactor: rename useQueryStorage to queryStoragePersister
The file exports a factory function, not a React hook. Rename to
accurately reflect its contents and update all imports.
